server/requirements.txt
﻿aiosqlite==0.20.0
alembic==1.14.0
annotated-types==0.7.0
anyio==4.13.0
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==5.0.0
bidict==0.23.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
cryptography==46.0.6
ecdsa==0.19.2
fastapi==0.115.8
google-auth==2.49.1
google-auth-oauthlib==1.3.0
greenlet==3.3.2
gspread==6.2.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.2
idna==3.11
loguru==0.7.2
Mako==1.3.10
MarkupSafe==3.0.3
oauthlib==3.3.1
orjson==3.10.12
passlib==1.7.4
Pillow==10.4.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.7.1
pydantic_core==2.41.5
pyotp==2.9.0
python-dotenv==1.0.1
python-engineio==4.13.1
python-jose==3.3.0
python-multipart==0.0.20
python-socketio==5.11.4
PyYAML==6.0.3
redis==5.2.0
requests==2.33.0
requests-oauthlib==2.0.0
rsa==4.9.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.45.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.30.6
watchfiles==1.1.1
websockets==16.0
win32_setctime==1.2.0
wsproto==1.3.2
